+++ src/Cocorico/UserBundle/Form/Type/ProfileBankAccountFormType.php
@@ -40,7 +40,6 @@ class ProfileBankAccountFormType extends AbstractType
                 'birthday',
                 array(
                     'label' => 'form.user.birthday',
-//                    'format' => 'dd MMMM yyyy',
                     'widget' => "choice",
                     'years' => range(date('Y') - 18, date('Y') - 100),
                     'required' => true
@@ -87,7 +86,6 @@ class ProfileBankAccountFormType extends AbstractType
                 array(
                     'label' => 'form.user.bank_owner_name',
                     'required' => true,
-//                    'disabled' => $bankEditionDisabled
                 )
             )
             ->add(